выделены сущности из MovieQuizViewController
