+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct QuizQuestion {
+  let image: String
+  let text: String
+  let correctAnswer: Bool
+}
